backend/pyproject.toml
[project]
name = "vedi-backend"
version = "0.1.0"
description = "Vedic Astrology Backend API"
authors = [
    {name = "tapiocaboy", email = "kalanamith@outlook.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pyswisseph>=2.10.3",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "pydantic>=2.5.0",
    "python-dateutil>=2.8.2",
    "pytz>=2024.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]

[tool.poetry]
packages = [{include = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

